resource-detector-aws
- chore(resource-detector-aws)!: merge Sync and async resources
  - removed `awsBeanstalkDetectorSync` -> use `awsBeanstalkDetector`
  - removed `awsEc2DetectorSync` -> use `awsEc2Detector`
  - removed `awsEcsDetectorSync` -> use `awsEcsDetector`
  - removed `awsEksDetectorSync` -> use `awsEksDetector`
  - removed `awsLambdaDetectorSync` -> use `awsLambdaDetector`
  - no longer exporting the detector classes: `AwsBeanstalkDetector`, `AwsEcsDetector`, `AwsEksDetector`, `AwsLambdaDetector`,
