+++ detectors/node/opentelemetry-resource-detector-aws/package.json
@@ -27,7 +27,7 @@
   "author": "OpenTelemetry Authors",
   "license": "Apache-2.0",
   "engines": {
-    "node": ">=14"
+    "node": "^18.19.0 || >=20.6.0"
   },
   "files": [
     "build/src/**/*.js",
@@ -44,7 +44,7 @@
     "@opentelemetry/api": "^1.0.0",
     "@opentelemetry/contrib-test-utils": "^0.45.1",
     "@opentelemetry/instrumentation-fs": "^0.19.1",
-    "@opentelemetry/instrumentation-http": "^0.57.2",
+    "@opentelemetry/instrumentation-http": "^0.200.0-dev.0",
     "@types/mocha": "10.0.10",
     "@types/node": "18.18.14",
     "@types/sinon": "17.0.4",
@@ -58,8 +58,8 @@
     "@opentelemetry/api": "^1.0.0"
   },
   "dependencies": {
-    "@opentelemetry/core": "^1.0.0",
-    "@opentelemetry/resources": "^1.10.0",
+    "@opentelemetry/core": "^2.0.0-dev.0",
+    "@opentelemetry/resources": "^2.0.0-dev.0",
     "@opentelemetry/semantic-conventions": "^1.27.0"
   },
   "homepage": "https://github.com/open-telemetry/opentelemetry-js-contrib/tree/main/detectors/node/opentelemetry-resource-detector-aws#readme",
