+++ detectors/node/opentelemetry-resource-detector-aws/src/detectors/AwsBeanstalkDetector.ts
@@ -14,13 +14,28 @@
  * limitations under the License.
  */
 
-import {
-  Detector,
-  IResource,
-  ResourceDetectionConfig,
-} from '@opentelemetry/resources';
+import { context } from '@opentelemetry/api';
+import { suppressTracing } from '@opentelemetry/core';
 
-import { awsBeanstalkDetectorSync } from './AwsBeanstalkDetectorSync';
+import {
+  ResourceDetector,
+  DetectedResource,
+  DetectedResourceAttributes,
+} from '@opentelemetry/resources';
+import {
+  ATTR_SERVICE_NAME,
+  ATTR_SERVICE_VERSION,
+} from '@opentelemetry/semantic-conventions';
+import {
+  ATTR_CLOUD_PROVIDER,
+  ATTR_CLOUD_PLATFORM,
+  ATTR_SERVICE_NAMESPACE,
+  ATTR_SERVICE_INSTANCE_ID,
+  CLOUD_PROVIDER_VALUE_AWS,
+  CLOUD_PLATFORM_VALUE_AWS_ELASTIC_BEANSTALK,
+} from '../semconv';
+import * as fs from 'fs';
+import * as util from 'util';
 
 /**
  * The AwsBeanstalkDetector can be used to detect if a process is running in AWS Elastic
@@ -29,13 +44,66 @@ import { awsBeanstalkDetectorSync } from './AwsBeanstalkDetectorSync';
  *
  * See https://docs.amazonaws.cn/en_us/xray/latest/devguide/xray-guide.pdf
  * for more details about detecting information of Elastic Beanstalk plugins
- *
- * @deprecated Use {@link AwsBeanstalkDetectorSync} class instead.
  */
 
-export class AwsBeanstalkDetector implements Detector {
-  detect(config?: ResourceDetectionConfig): Promise<IResource> {
-    return Promise.resolve(awsBeanstalkDetectorSync.detect(config));
+const DEFAULT_BEANSTALK_CONF_PATH =
+  '/var/elasticbeanstalk/xray/environment.conf';
+const WIN_OS_BEANSTALK_CONF_PATH =
+  'C:\\Program Files\\Amazon\\XRay\\environment.conf';
+
+export class AwsBeanstalkDetector implements ResourceDetector {
+  BEANSTALK_CONF_PATH: string;
+  private static readFileAsync = util.promisify(fs.readFile);
+  private static fileAccessAsync = util.promisify(fs.access);
+
+  constructor() {
+    if (process.platform === 'win32') {
+      this.BEANSTALK_CONF_PATH = WIN_OS_BEANSTALK_CONF_PATH;
+    } else {
+      this.BEANSTALK_CONF_PATH = DEFAULT_BEANSTALK_CONF_PATH;
+    }
+  }
+
+  detect(): DetectedResource {
+    const attributes = context.with(suppressTracing(context.active()), () =>
+      this._getAttributes()
+    );
+    return { attributes };
+  }
+
+  /**
+   * Async resource attributes for AWS Beanstalk configuration read from file.
+   */
+  _getAttributes(): DetectedResourceAttributes {
+    const parsedDataP = AwsBeanstalkDetector.fileAccessAsync(
+      this.BEANSTALK_CONF_PATH,
+      fs.constants.R_OK
+    )
+      .then(() =>
+        AwsBeanstalkDetector.readFileAsync(this.BEANSTALK_CONF_PATH, 'utf8')
+      )
+      .then(rawData => {
+        return JSON.parse(rawData);
+      });
+
+    return {
+      [ATTR_CLOUD_PROVIDER]: parsedDataP.then(() => CLOUD_PROVIDER_VALUE_AWS),
+      [ATTR_CLOUD_PLATFORM]: parsedDataP.then(
+        () => CLOUD_PLATFORM_VALUE_AWS_ELASTIC_BEANSTALK
+      ),
+      [ATTR_SERVICE_NAME]: parsedDataP.then(
+        () => CLOUD_PLATFORM_VALUE_AWS_ELASTIC_BEANSTALK
+      ),
+      [ATTR_SERVICE_NAMESPACE]: parsedDataP.then(
+        parsedData => parsedData.environment_name
+      ),
+      [ATTR_SERVICE_VERSION]: parsedDataP.then(
+        parsedData => parsedData.version_label
+      ),
+      [ATTR_SERVICE_INSTANCE_ID]: parsedDataP.then(
+        parsedData => parsedData.deployment_id
+      ),
+    };
   }
 }
