+++ detectors/node/opentelemetry-resource-detector-aws/src/detectors/index.ts
@@ -14,22 +14,8 @@
  * limitations under the License.
  */
 
-export {
-  AwsBeanstalkDetector,
-  awsBeanstalkDetector,
-} from './AwsBeanstalkDetector';
-export {
-  AwsBeanstalkDetectorSync,
-  awsBeanstalkDetectorSync,
-} from './AwsBeanstalkDetectorSync';
+export { awsBeanstalkDetector } from './AwsBeanstalkDetector';
 export { awsEc2Detector } from './AwsEc2Detector';
-export { awsEc2DetectorSync } from './AwsEc2DetectorSync';
-export { AwsEcsDetector, awsEcsDetector } from './AwsEcsDetector';
-export { AwsEcsDetectorSync, awsEcsDetectorSync } from './AwsEcsDetectorSync';
-export { AwsEksDetector, awsEksDetector } from './AwsEksDetector';
-export { AwsEksDetectorSync, awsEksDetectorSync } from './AwsEksDetectorSync';
-export { AwsLambdaDetector, awsLambdaDetector } from './AwsLambdaDetector';
-export {
-  AwsLambdaDetectorSync,
-  awsLambdaDetectorSync,
-} from './AwsLambdaDetectorSync';
+export { awsEcsDetector } from './AwsEcsDetector';
+export { awsEksDetector } from './AwsEksDetector';
+export { awsLambdaDetector } from './AwsLambdaDetector';
