+++ detectors/node/opentelemetry-resource-detector-aws/test/detectors/AwsBeanstalkDetector.test.ts
@@ -16,12 +16,14 @@
 
 import * as assert from 'assert';
 import * as sinon from 'sinon';
-import { awsBeanstalkDetector, AwsBeanstalkDetectorSync } from '../../src';
 import {
   assertEmptyResource,
   assertServiceResource,
 } from '@opentelemetry/contrib-test-utils';
+import { detectResources } from '@opentelemetry/resources';
 import { CLOUDPLATFORMVALUES_AWS_ELASTIC_BEANSTALK } from '@opentelemetry/semantic-conventions';
+import { awsBeanstalkDetector } from '../../src';
+import { AwsBeanstalkDetector } from '../../src/detectors/AwsBeanstalkDetector';
 
 describe('BeanstalkResourceDetector', () => {
   const err = new Error('failed to read config file');
@@ -45,14 +47,14 @@ describe('BeanstalkResourceDetector', () => {
 
   it('should successfully return resource data', async () => {
     fileStub = sinon
-      .stub(AwsBeanstalkDetectorSync, 'fileAccessAsync' as any)
+      .stub(AwsBeanstalkDetector, 'fileAccessAsync' as any)
       .resolves();
     readStub = sinon
-      .stub(AwsBeanstalkDetectorSync, 'readFileAsync' as any)
+      .stub(AwsBeanstalkDetector, 'readFileAsync' as any)
       .resolves(JSON.stringify(data));
     sinon.stub(JSON, 'parse').returns(data);
 
-    const resource = await awsBeanstalkDetector.detect();
+    const resource = detectResources({ detectors: [awsBeanstalkDetector] });
     await resource.waitForAsyncAttributes?.();
 
     sinon.assert.calledOnce(fileStub);
@@ -68,14 +70,14 @@ describe('BeanstalkResourceDetector', () => {
 
   it('should successfully return resource data with noise', async () => {
     fileStub = sinon
-      .stub(AwsBeanstalkDetectorSync, 'fileAccessAsync' as any)
+      .stub(AwsBeanstalkDetector, 'fileAccessAsync' as any)
       .resolves();
     readStub = sinon
-      .stub(AwsBeanstalkDetectorSync, 'readFileAsync' as any)
+      .stub(AwsBeanstalkDetector, 'readFileAsync' as any)
       .resolves(JSON.stringify(noisyData));
     sinon.stub(JSON, 'parse').returns(noisyData);
 
-    const resource = await awsBeanstalkDetector.detect();
+    const resource = detectResources({ detectors: [awsBeanstalkDetector] });
     await resource.waitForAsyncAttributes?.();
 
     sinon.assert.calledOnce(fileStub);
@@ -91,13 +93,13 @@ describe('BeanstalkResourceDetector', () => {
 
   it('should return empty resource when failing to read file', async () => {
     fileStub = sinon
-      .stub(AwsBeanstalkDetectorSync, 'fileAccessAsync' as any)
+      .stub(AwsBeanstalkDetector, 'fileAccessAsync' as any)
       .resolves();
     readStub = sinon
-      .stub(AwsBeanstalkDetectorSync, 'readFileAsync' as any)
+      .stub(AwsBeanstalkDetector, 'readFileAsync' as any)
       .rejects(err);
 
-    const resource = await awsBeanstalkDetector.detect();
+    const resource = detectResources({ detectors: [awsBeanstalkDetector] });
     await resource.waitForAsyncAttributes?.();
 
     sinon.assert.calledOnce(fileStub);
@@ -108,13 +110,13 @@ describe('BeanstalkResourceDetector', () => {
 
   it('should return empty resource when config file does not exist', async () => {
     fileStub = sinon
-      .stub(AwsBeanstalkDetectorSync, 'fileAccessAsync' as any)
+      .stub(AwsBeanstalkDetector, 'fileAccessAsync' as any)
       .rejects(err);
     readStub = sinon
-      .stub(AwsBeanstalkDetectorSync, 'readFileAsync' as any)
+      .stub(AwsBeanstalkDetector, 'readFileAsync' as any)
       .resolves(JSON.stringify(data));
 
-    const resource = await awsBeanstalkDetector.detect();
+    const resource = detectResources({ detectors: [awsBeanstalkDetector] });
     await resource.waitForAsyncAttributes?.();
 
     sinon.assert.calledOnce(fileStub);
